+++ test/baselines/adapter_JS_JsonUI_ESLint_TypeChecking_Spaces_SingleQuotes_Apache-2.0/main.js
@@ -1,7 +1,7 @@
 'use strict';
 
 /*
- * Created with @iobroker/create-adapter v2.6.0
+ * Created with @iobroker/create-adapter v2.6.1
  */
 
 // The adapter-core module gives you access to the core ioBroker functions