+++ test/baselines/adapter_TS_ESLint_Tabs_DoubleQuotes_MIT/src/main.ts
@@ -1,5 +1,5 @@
 /*
- * Created with @iobroker/create-adapter v2.6.0
+ * Created with @iobroker/create-adapter v2.6.1
  */
 
 // The adapter-core module gives you access to the core ioBroker functions